tests: create isolated test file
